move "Warn about duplicates on import" preference option
